TP AOP
- Utilisation des annotations pour les aspect et les pointcut
- Modification du systeme d'enregistrement des pizzas

+++ Jeffrey-pizzeria-app/Jeffrey-pizzeria-model/src/main/java/fr/pizzeria/model/Livreur.java
@@ -20,7 +20,7 @@ public class Livreur {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private byte id;
+	private Integer id;
 	private String nom;
 	private String prenom;
 
@@ -31,11 +31,11 @@ public class Livreur {
 	public Livreur(){
 	}
 
-	public byte getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(byte id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
